+++ src/snl/snl/decorators/SNLTruthTable.h
@@ -18,7 +18,7 @@ namespace naja { namespace SNL {
 class SNLTruthTable {
   public:
     SNLTruthTable(): size_(0xFF) {}
-    explicit SNLTruthTable(uint8_t size, uint64_t bits): size_(size), bits_(bits) {
+    explicit SNLTruthTable(uint32_t size, uint64_t bits): size_(size), bits_(bits) {
       if (size > 6) {
         throw SNLException("Size out of range (max=6)");
       }
@@ -42,17 +42,17 @@ class SNLTruthTable {
       return size_ != 0xFF;
     }
 
-    SNLTruthTable getReducedWithConstant(uint8_t index, bool constant) const {
+    SNLTruthTable getReducedWithConstant(uint32_t index, bool constant) const {
       if (index > 6) {
         throw SNLException("Index out of range (max=6)");
       }
       if (size_ == 0) {
         return *this;
       }
-      uint8_t n = (uint8_t)(1 << size_);
+      uint32_t n = 1 << size_;
       uint64_t reducedBits = 0;
-      uint8_t bitPos = 0;
-      for (uint8_t i = 0; i < n; ++i) {
+      uint32_t bitPos = 0;
+      for (uint32_t i = 0; i < n; ++i) {
         if (((i >> index) & 1) == constant) {
             reducedBits |= ((bits_ >> i) & 1) << bitPos;
             ++bitPos;
@@ -69,21 +69,21 @@ class SNLTruthTable {
     }
 
     // Function to check if an input has no influence on the output
-    bool hasNoInfluence(uint8_t variableIndex) const {
+    bool hasNoInfluence(uint32_t variableIndex) const {
       SNLTruthTable tableWithZero = getReducedWithConstant(variableIndex, false);
       SNLTruthTable tableWithOne = getReducedWithConstant(variableIndex, true);
       return tableWithZero == tableWithOne;
     }    
 
     // Function to remove a variable from the truth table
-    SNLTruthTable removeVariable(uint8_t variableIndex) const {
+    SNLTruthTable removeVariable(uint32_t variableIndex) const {
       if (variableIndex > size_) {
         throw SNLException("Index out of range");
       }
       SNLTruthTable reducedTruthTable(size_-1, 0);
-      for (uint8_t i = 0; i < (1 << size_); ++i) {
+      for (uint32_t i = 0; i < (1 << size_); ++i) {
         if (((i >> variableIndex) & 1) == 0) {
-            int newIdx = ((i & ((1 << variableIndex) - 1)) | ((i >> 1) & (~((1 << variableIndex) - 1))));
+            uint32_t newIdx = ((i & ((1 << variableIndex) - 1)) | ((i >> 1) & (~((1 << variableIndex) - 1))));
             if (((bits_ >> i) & 1) == 1) {
               reducedTruthTable.bits_ |= (1 << newIdx);
             }
@@ -114,7 +114,7 @@ class SNLTruthTable {
       return bits_;
     }
   private:
-    uint8_t   size_ {0};
+    uint32_t  size_ {0};
     uint64_t  bits_ {0};
 };
